Migrate separate SSO details flow into existing one
